Update M5Dial_Timezones_and_M5Echo.ino
Added functionality to Display.sleep() and Display.awake(). Changed the way the text "TS" in the top of the Display is being cleared. Ended use of partially clear the Display. Added functionality to Switch display off and ON by Display touches. Added that no beep command will be send to the Atom Echo device when the Display is off.